+++ shared/core/src/declarations/base/index.ts
@@ -1,4 +1,25 @@
-export * from './core';
+/**
+ * 获取方法中参数的类型
+ */
+export type ParametersType<T> = T extends (args: infer R) => any ? R : any;
+
+/**
+ * 数据字典定义
+ */
+export type Dictionary = {
+  ordinal: number;
+  name: string;
+  value: string;
+  label: string;
+};
+
+export interface Validation {
+  validate: () => Promise<boolean>;
+}
+
+export type { QTableProps, QNotifyCreateOptions } from 'quasar';
+
+export type { SweetAlertIcon, SweetAlertResult } from 'sweetalert2';
 
 export type {
   RawAxiosRequestConfig,
@@ -13,7 +34,3 @@ export type {
   Canceler,
   ParamsSerializerOptions,
 } from 'axios';
-
-export type { QTableProps, QNotifyCreateOptions } from 'quasar';
-
-export type { SweetAlertIcon, SweetAlertResult } from 'sweetalert2';