+++ shared/core/src/declarations/components/axios.ts
@@ -5,7 +5,7 @@ import type {
   AxiosResponse,
   AxiosError,
   InternalAxiosRequestConfig,
-} from '../base';
+} from '@/declarations';
 
 import type { HttpResult, AxiosHttpResult } from '../modules';
 
@@ -27,19 +27,16 @@ export interface RequestOptions {
   withToken?: boolean;
 }
 
-export interface AxiosTransform {
+export interface AxiosHook {
   /**
    * @description: 请求前处理配置
    */
-  beforeRequestHook?: (
-    config: RawAxiosRequestConfig,
-    options: RequestOptions,
-  ) => RawAxiosRequestConfig;
+  onRequestHook?: (config: RawAxiosRequestConfig, options: RequestOptions) => RawAxiosRequestConfig;
 
   /**
    * @description: 请求成功处理
    */
-  transformRequestHook?: <D = unknown>(
+  onResponseSuccessHook?: <D = unknown>(
     response: AxiosResponse<HttpResult<D>>,
     options?: RequestOptions,
   ) => AxiosHttpResult<D>;
@@ -47,7 +44,7 @@ export interface AxiosTransform {
   /**
    * @description: 请求失败处理
    */
-  requestCatchHook?: <D = unknown>(error: AxiosError, options?: RequestOptions) => HttpResult<D>;
+  onResponseErrorHook?: <D = unknown>(error: AxiosError, options?: RequestOptions) => HttpResult<D>;
 
   /**
    * @description: 请求之前的拦截器
@@ -62,27 +59,29 @@ export interface AxiosTransform {
   /**
    * @description: 请求之前的拦截器错误处理
    */
-  requestInterceptorsCatch: (axiosInstance: AxiosInstance, error: AxiosError) => Promise<any>;
+  requestInterceptorsError: (axiosInstance: AxiosInstance, error: AxiosError) => Promise<any>;
 
   /**
    * @description: 请求之后的拦截器错误处理
    */
-  responseInterceptorsCatch: (axiosInstance: AxiosInstance, error: AxiosError) => Promise<any>;
+  responseInterceptorsError: (axiosInstance: AxiosInstance, error: AxiosError) => Promise<any>;
 }
 
 export interface AxiosConfig extends RawAxiosRequestConfig {
   authenticationScheme?: string;
-  transform?: AxiosTransform;
+  hook?: AxiosHook;
   requestOptions?: RequestOptions;
 }
 
-export interface Policy {
-  headers: RawAxiosRequestHeaders;
+interface AxiosParameterConverter {
   dataConvert: (params: Record<string, any>) => any;
 }
 
-export interface AxiosRequestPolicy {
+export interface AxiosHeaderStrategy extends AxiosParameterConverter {
+  headers: RawAxiosRequestHeaders;
+}
+
+export interface AxiosRequestStrategy extends AxiosParameterConverter {
   config: RawAxiosRequestConfig;
   options: RequestOptions;
-  dataConvert: (params: Record<string, any>) => any;
 }