+++ shared/framework-kernel/src/lib/hooks/index.ts
@@ -1,4 +1,5 @@
 import useEditFinish from './useEditFinish';
+import useDeviceAuthorize from './useDeviceAuthorize';
 import useMenuForQuasar from './useMenuForQuasar';
 import useMenuForVuetify from './useMenuForVuetify';
 import usePasskey from './usePasskey';
@@ -7,6 +8,7 @@ import useSystemTheme from './useSystemTheme';
 
 export {
   useEditFinish,
+  useDeviceAuthorize,
   usePasskey,
   useMenuForQuasar,
   useMenuForVuetify,